Color Code and Saving Text
I still cannot figure out how to save text to input field and stay there after reloading the page. I also found a way to write a function for keeping the colors on the input fields based on time. However, there is a problem with the var moment() and it is return an error.