Color Coding
Finally fixed my bug! I left the html with the past, present, and future in the class and it prevented the function from labeling the text boxes currently with the time. I also used the moment.js documentation online to convert to day.js to call the currentTime from the displayTime function.